extract common helpers for testing
